Moving ConventionsTrait to Core
