+++ src/Core/ConventionsTrait.php
@@ -12,7 +12,7 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Utility;
+namespace Cake\Core;
 
 use Cake\Core\App;
 use Cake\Core\Plugin;
