+++ src/Shell/BakeShell.php
@@ -20,7 +20,7 @@ use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Datasource\ConnectionManager;
 use Cake\Model\Model;
-use Cake\Utility\ConventionsTrait;
+use Cake\Core\ConventionsTrait;
 use Cake\Utility\Inflector;
 
 /**
