+++ src/Shell/Task/TemplateTask.php
@@ -17,7 +17,7 @@ namespace Cake\Shell\Task;
 use Cake\Console\Shell;
 use Cake\Core\App;
 use Cake\Core\Plugin;
-use Cake\Utility\ConventionsTrait;
+use Cake\Core\ConventionsTrait;
 use Cake\Utility\Folder;
 use Cake\View\ViewVarsTrait;
